test: adding Jasmine tests in FeedReader application
In this project the testing is the main purpose. So, using Jasmine for JavaScript
we utilise tests against the application that feeds stuff from other sites.
With these tests we check the feeds to be valid and we check the menu functionality
of the UI. Also we check the feeding process to work properly.

This is a testing project of Udacity Front-End Nano Degree